% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_match_lr_pairs}
\alias{rust_match_lr_pairs}
\title{Match overexpressed genes against CellChatDB ligand-receptor pairs.}
\usage{
rust_match_lr_pairs(overexpressed, lr_ligands, lr_receptors, lr_names)
}
\arguments{
\item{overexpressed}{Character vector of overexpressed gene names.}

\item{lr_ligands}{Character vector of ligands from CellChatDB.}

\item{lr_receptors}{Character vector of receptors from CellChatDB.}

\item{lr_names}{Character vector of interaction names from CellChatDB.}
}
\value{
Character vector of matched interaction names.
}
\description{
Match overexpressed genes against CellChatDB ligand-receptor pairs.
}
