% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_subset_genes}
\alias{rust_subset_genes}
\title{Filter expression matrix genes to only those present in CellChatDB.}
\usage{
rust_subset_genes(expr_genes, db_genes)
}
\arguments{
\item{expr_genes}{Character vector of gene names from expression matrix.}

\item{db_genes}{Character vector of gene names from CellChatDB.}
}
\value{
Character vector of genes present in both inputs.
}
\description{
Filter expression matrix genes to only those present in CellChatDB.
}
\section{Complexity}{
\itemize{
\item Time:  O(n + m) where n = db_genes, m = expr_genes
\item Space: O(n) for HashSet
\item R equivalent (\%\%in\%\%): O(n x m)
}
}

\examples{
rust_subset_genes(
  c("TGFB1", "VEGFA", "GAPDH"),
  c("TGFB1", "VEGFA", "MYC")
)
}
